pyproject.toml
[project]
name = "github-repo-summarizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncio>=4.0.0",
    "fastmcp>=2.14.5",
    "google-generativeai>=0.8.6",
    "mcp[cli]>=1.26.0",
    "pathlib>=1.0.1",
    "requests>=2.32.5",
]
